add event listeners, dimension filter update demo remove deprecated bar chart

+++ src/com/crossfilter/client/SingleGroup.java
@@ -19,7 +19,7 @@ public class SingleGroup<T, K> extends JavaScriptObject {
 		return this.value();
     }-*/;
 
-    public final native double getValueInt() /*-{
+    public final native int getValueInt() /*-{
 		return this.value();
     }-*/;
 }
